Moved shooting functionality to new Gun script
