+++ Assets/Prefabs/Gun.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e9e02d255324a0749a642801ae1cdeaa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
