+++ Assets/Prefabs/Player.prefab
@@ -111,6 +111,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 8211571826163515460}
+  - {fileID: 8211571826770355695}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -127,6 +128,82 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 6
-  ProjectilePrefab: {fileID: 8274613288791820710, guid: 6b1123ecd6ee2fc4590b836339b04633, type: 3}
-  bulletSpeed: 6
-  fireRate: 4
+  gun: {fileID: 8211571826770355692}
+  moveDirection: {x: 0, y: 0, z: 0}
+--- !u!1001 &4753104144659895456
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 8211571826743374638}
+    m_Modifications:
+    - target: {fileID: 3459668417817309516, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: player
+      value: 
+      objectReference: {fileID: 8211571826743374633}
+    - target: {fileID: 3459668417817309518, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_Name
+      value: Gun
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+--- !u!4 &8211571826770355695 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 3459668417817309519, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+  m_PrefabInstance: {fileID: 4753104144659895456}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &8211571826770355692 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3459668417817309516, guid: e9e02d255324a0749a642801ae1cdeaa, type: 3}
+  m_PrefabInstance: {fileID: 4753104144659895456}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7fc73ad1361d16e45aca23871d0bc1ee, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
