+++ Assets/Scripts/Gun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7fc73ad1361d16e45aca23871d0bc1ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
