+++ Assets/Scripts/Player.cs
@@ -2,22 +2,16 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    [SerializeField]
-    float speed = 6.0f;
+    public float speed = 6.0f;
 
     [SerializeField]
-    GameObject ProjectilePrefab;
+    Gun gun;
 
-    [SerializeField]
-    float bulletSpeed = 10.0f;    
-    [SerializeField]
-    float fireRate = 1.0f;  //  Bullets per second
 
-    //  Private:
-    Vector3 moveDirection;
+    //  Directions:
+    public Vector3 moveDirection;
     Vector3 shootDirection;
 
-    float cooldown;
 
     void Start()
     {
@@ -30,14 +24,7 @@ public class Player : MonoBehaviour
         CheckInput();
     }
 
-    void SpawnProjectile(Vector3 direction) 
-        //  I'm using Vector3 here to leave space for possible diagonal shooting in future
-    {
-        GameObject proj = Instantiate(ProjectilePrefab, transform.position, Quaternion.identity);
-        Projectile bullet = proj.GetComponent<Projectile>();
-        bullet.SetSpeed(direction * bulletSpeed + moveDirection * speed);
 
-    }
 
     void CheckInput()
     {
@@ -60,40 +47,9 @@ public class Player : MonoBehaviour
 
         float sx = Input.GetAxis("ShootHorizontal"); //  Get input
         float sy = Input.GetAxis("ShootVertical");
-
-
-
-        cooldown -= Time.deltaTime; //  Tick cooldown timer
-
-        if (cooldown < 0)
-            cooldown = 0;   //  Prevent negative values
-
-
         if (sx != 0 || sy != 0) //  Check for shooting input
         {
-            if (sy > 0) //  Set shooting direction vector
-            {
-                shootDirection = new Vector3(0.0f, 1.0f, 0.0f);
-            }
-            else if (sy < 0)
-            {
-                shootDirection = new Vector3(0.0f, -1.0f, 0.0f);
-            }
-            if (sx > 0)
-            {
-                shootDirection = new Vector3(1.0f, 0.0f, 0.0f);
-            }
-            else if(sx < 0)
-            {
-                shootDirection = new Vector3(-1.0f, 0.0f, 0.0f);
-            }
-            //Debug.Log(shootDirection);
-
-            if(cooldown == 0)
-            {
-                cooldown += 1/fireRate;             //  Reset cooldown
-                SpawnProjectile(shootDirection);    //  Shoot
-            }
+            gun.Shoot(sx, sy);
         }
 
     }
